+++ src/index.ts
@@ -1,4 +1,3 @@
-/// <reference path="./types/optional-peer-deps.d.ts" />
 export * from './datagrid';
 export * from './drag-drop';
 export * from './pagination';