+++ src/types/optional-peer-deps.d.ts
@@ -1,3 +1,0 @@
-declare module 'jspdf';
-declare module 'jspdf/dist/jspdf.umd.min.js';
-declare module 'jspdf-autotable';